Reduce usage of util files
